+++ R/np.qregression.R
@@ -22,7 +22,8 @@ npqreg <-
     }
   }
 
-.npqreg.fit.control.names <- c("data", "newdata", "exdat", "tau", "gradients", "tol", "small", "itmax")
+.npqreg.fit.control.names <- c("data", "newdata", "exdat", "tau", "gradients", "tol", "small", "itmax",
+                             ".np_conditional_cat_se_demand")
 .npqreg.removed.solver.controls <- c("ftol",
                                      "lbc.dir", "dfc.dir", "cfac.dir", "initc.dir",
                                      "lbd.dir", "hbd.dir", "dfac.dir", "initd.dir")
@@ -170,7 +171,8 @@ npqreg <-
 }
 
 .npqreg_strip_fit_controls_from_bw_call <- function(call) {
-  for (nm in c("tau", "gradients", "tol", "small", "itmax", "newdata", "exdat")) {
+  for (nm in c("tau", "gradients", "tol", "small", "itmax", "newdata", "exdat",
+               ".np_conditional_cat_se_demand")) {
     if (nm %in% names(call))
       call[[nm]] <- NULL
   }
@@ -188,10 +190,16 @@ npqreg <-
                                                     small = 1.490116e-05,
                                                     itmax = 10000L,
                                                     cdf.cache = NULL,
-                                                    lp.first.se.demand = NULL) {
+                                                    lp.first.se.demand = NULL,
+                                                    cat.se.demand = NULL) {
   xdat <- toFrame(xdat)
   ydat <- toFrame(ydat)
   exdat <- toFrame(exdat)
+  cat.se.demand <- .np_conditional_cat_se_demand(
+    cat.se.demand, bws$xnuno + bws$xnord)
+  qclamp <- .npqreg_quantile_clamp(quantile)
+  if (all(!is.na(qclamp) & qclamp != "none"))
+    cat.se.demand[] <- FALSE
   quantile <- as.double(quantile)
   gradients <- npValidateScalarLogical(gradients, "gradients")
 
@@ -225,7 +233,8 @@ npqreg <-
     cdf = TRUE,
     gradients = gradients,
     categorical.effects = !glp.categorical.effects,
-    lp.first.se.demand = lp.first.se.demand
+    lp.first.se.demand = lp.first.se.demand,
+    cat.se.demand = if (glp.categorical.effects) FALSE else cat.se.demand
   )
   dens.obj <- .np_conditional_eval_selected(
     bws = bws,
@@ -638,7 +647,12 @@ npqreg.condbandwidth <-
            ...){
 
     fit.start <- proc.time()[3]
-    fit.dots <- .npqreg_fit_dots(list(...))
+    fit.dots <- list(...)
+    cat.se.demand <- .np_conditional_cat_se_demand(
+      fit.dots[[".np_conditional_cat_se_demand", exact = TRUE]],
+      bws$xnuno + bws$xnord)
+    fit.dots[[".np_conditional_cat_se_demand"]] <- NULL
+    fit.dots <- .npqreg_fit_dots(fit.dots)
     if (length(fit.dots))
       stop(sprintf("unused npqreg fit argument '%s'", names(fit.dots)[1L]))
     gradients <- npValidateScalarLogical(gradients, "gradients")
@@ -766,7 +780,8 @@ npqreg.condbandwidth <-
         tol = tol,
         small = small,
         itmax = itmax,
-        cdf.cache = cdf.cache
+        cdf.cache = cdf.cache,
+        cat.se.demand = cat.se.demand
       )
       qdelta <- .npqreg_mark_clamped_delta(qdelta, qclamp)
       list(